finished admin forms with spring mvc

+++ src/main/java/com/web/UserController.java
@@ -1,15 +1,16 @@
 package com.web;
 
+import com.domain.Reservation;
 import com.domain.User;
+import com.service.ReservationService;
 import com.service.UserService;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Controller;
-import org.springframework.web.bind.annotation.PathVariable;
-import org.springframework.web.bind.annotation.RequestBody;
-import org.springframework.web.bind.annotation.RequestMapping;
-import org.springframework.web.bind.annotation.ResponseBody;
+import org.springframework.web.bind.annotation.*;
 import org.springframework.web.servlet.ModelAndView;
+import org.springframework.web.servlet.view.RedirectView;
 
+import javax.servlet.http.HttpServletResponse;
 import javax.xml.bind.DatatypeConverter;
 import java.security.MessageDigest;
 import java.util.List;
@@ -20,10 +21,12 @@ public class UserController
     @Autowired
     private UserService userService;
 
+    @Autowired
+    private ReservationService reservationService;
+
     @RequestMapping("/admin/addUser")
-    public @ResponseBody ModelAndView addUser(@RequestBody User user) throws Exception
+    public ModelAndView addUser(@ModelAttribute User user) throws Exception
     {
-        ModelAndView mav = new ModelAndView("addUser");
         if(user != null && user.getLogin() != null
                 && user.getPassword() != null && user.getEmail() != null)
         {
@@ -31,43 +34,63 @@ public class UserController
             digest.reset();
             byte[] hash = digest.digest(user.getPassword().getBytes("UTF-8"));
             String passwordHash = DatatypeConverter.printHexBinary(hash);
-            List<User> userList = userService.getUserAll();
-            boolean userValid = true;
             user.setPassword(passwordHash);
-            for(User u : userList)
-            {
-                if(u.getLogin().equals(user.getLogin()) || u.getEmail().equals(user.getEmail()))
-                {
-                    userValid = false;
-                }
-            }
-            if(userValid)
-            {
-                userService.create(user);
-                mav.addObject("userJson", user);
-            }
+            userService.create(user);
+            return new ModelAndView(new RedirectView("/cinema/admin/addUser"));
         }
-        return mav;
+        return new ModelAndView("addUser");
     }
 
 //    @RequestMapping("/admin/updateUser")
 
-    @RequestMapping("/admin/deleteUser/{userId}")
-    public @ResponseBody ModelAndView deleteUser(@PathVariable int userId) throws Exception
+    @RequestMapping("/admin/deleteUser")
+    public ModelAndView deleteUser(@ModelAttribute User user, HttpServletResponse response) throws Exception
     {
-        List<User> userList = userService.getUserAll();
-        User user = userService.getUserById(userId);
-        if(user != null)
+        try
         {
-            userService.delete(user);
+            if(user.getUserId() != null && user.getUserId() != 0)
+            {
+                user = userService.getUserById(user.getUserId());
+                if(user != null)
+                {
+                    userService.delete(user);
+                    return new ModelAndView(new RedirectView("deleteUser"));
+                }
+            }
+            List<User> userList = userService.getUserAll();
+            response.setStatus(HttpServletResponse.SC_OK);
+            return new ModelAndView("deleteUser", "userList", userList);
         }
-        return new ModelAndView("deleteUser", "userListJson", userList);
+        catch(Exception e)
+        {
+            response.setStatus(HttpServletResponse.SC_INTERNAL_SERVER_ERROR);
+            e.printStackTrace();
+            return null;
+        }
     }
 
     @RequestMapping("/admin/userList")
-    public @ResponseBody ModelAndView listUsers() throws Exception
+    public ModelAndView listUsers() throws Exception
     {
         List<User> userList = userService.getUserAll();
-        return new ModelAndView("userList", "userListJson", userList);
+        return new ModelAndView("userList", "userList", userList);
+    }
+
+    @RequestMapping(value = "/admin/userCheck/{userId}", produces = "text/html; charset=UTF-8")
+    public @ResponseBody String userCheck(@PathVariable int userId) throws Exception
+    {
+        User user = userService.getUserById(userId);
+        if(user != null)
+        {
+            List<Reservation> reservationList = reservationService.getReservationAll();
+            for(Reservation r : reservationList)
+            {
+                if(r.getUser().equals(user))
+                {
+                    return "У пользователя есть брони. Сначала удалите бронь.";
+                }
+            }
+        }
+        return null;
     }
 }
